Add debug prints, fix a pan/zoom bug, made some text fields wider
Add debug statements to wheel_scroll_zoomx(), gototf_Return(), and
init/main code, to learn what is happening.

Make some of the text fields at the bottom of the window wider, to see
more of the text inside of them.

Fix a bug in wheel_scroll_zoomx() that only occurred after clicking
pan/zoom button, which changed the text of a field that
wheel_scroll_zoomx() used.