Added & use new function set_xlim_all, simplifed newxrange calc, etc.
Added new function set_xlim_all() to make it easier to keep the time
range of all graphs equal to each other.  Changed all places that
updated the time range to use this new function.

Made lastevtime initially None, not 0, and everywher lastevtime is
accessed, a check is made to see if it is None, and if so, it is
initialized to something reasonable.

Simplified the calculation of newxrange in function
wheel_scroll_zoomx.

Deleted white space at the end of lines that are not all blank.