Simplify newxrange calc in zoomtf_Return(), explicit globals, reformatting
Make all 'global' declarations complete according to current function
definitions.  It also declares all globals that are read-only inside
of a function, which isn't necessary in Python, but is harmless and
makes the fact of the globalness explicit.

Renamed global 'r' to much more descriptive 'xy_parsing_regex'.

Simplify newxrange calculation in zoomtf_Return() as it was done for
wheel_scroll_zoomx() in a recent commit.

Many small formatting changes for my readability.